Use an enum for the ThumnailFlavor.
Write some code to convert underscores to hyphens.  Dart enums cannot have variants that have
an '-' charcter.  We can use the '_' character instead.

There are two thumbnail flavors that have hyphens in their names.
These are "x-large", and "xx-large".  The enumeration variants are
"x_large", and "xx_large".  These variants do trigger a warning about constant identifier
names needing to be in camelCase.  Added an ignore_for_file for this rule to avoid this warning.