Obtain future in advance.
With a FutureBuilder, we need to obtain the future in advance, otherwise, whenever the FutureBuilder's parent
gets rebuilt, the async task gets restarted.

Fixed by creating _thumbnailFuture, and storing the future here in initState().

Also change the styling of the error so that the icon is bigger, and is colored red.